:recycle: move track changes service to builder state
