+++ libs/state/convs-mgr/wflows/main/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/providers/wflow.service';
-export * from './lib/services/track-changes.service'
 export * from './lib/providers/flow-editor-state.provider';
 export * from './lib/stores/whatsapp-flow.store'
